mongodb/requirements.txt
fastapi
pydantic
uvicorn
pymongo
python-jose
pytest
python-dotenv
python-multipart
motor
httpx
alembic
hypothesis
pytest-asyncio